Initial work on file list scrolling
If you get to the end of the listed files, and then attempt to scroll down,
the top most item(s) will fall off the list to make room for new items.

At the moment, there is no lower bound to the scrolling, meaning that you can
scroll off the bottom of the list and then will have to scroll back up the same
number of items.