Preparations for sorting files
File info is now loaded into a QueueList, then displayed. Still need to sort
the list before display. This means that we are now dynamically allocating
RAM based on the number of files on the SD card, which could be problematic
if there are a lot of files.

Also added checks preventing scrolling to invalid lines (empty entries, past
the end of the file list).