Sort file browser entries
The QueueList sorting modifications should work for any QueueList, but
have only been tested with the FileInfo struct.

The FileInfo sort order is not ideal, but it is passable. Numbers come
before letters, and letters are treated as case insensitive. It will
still have trouble with numerical series, meaning (1, 2, 20, 10) will
become (1, 10, 2, 20).