Add line wrapping for target file path
Line wrapping attempts to break at file system dividers ('/') or at
whitepsace ('\n' or ' '). Even with this though, we can still only
fit two lines of text. It also does not currently break long "words",
which is a problem if a single element in the file path is too long.