Make drawing the BSOD 40% faster.
Dropped draw time for BSOD from 231 ms to 131 ms by switching to
drawFastHLine.

Since the SPI transaction time still is a significant part of the
wait, I suspect I could make things even faster by updating the
library to allow pushing a buffer straight to the display, but I
don't think that is worth it. This is fast enough, IMO.