Slave firmware is feature-complete!
This should be stable enough for an initial release.

Note: the reason the allowed error is so hihg (20%) is because one
of my coils appears to have some damage (internal shorts) and is
~15% out from the rest. So for now, I just allow it, but I might
repair/replace the damaged coil and tighten up the allowed error.
The rest of the coils are within 1% of expected from what I see.