Make slaves 1-indexed & work around Teensy/GCC bug.
Slaves are now numbered starting from 1 in the Settings.ini file and in
all program outputs. This is now consistent with how channels are
presented, and is more intuitive for "normal" people.

There is a bug where malloc() and new both fail spectacularly with the
Teensy's version of GCC. This is visible when attempting to malloc()
more RAM than is available. Instead of returning a null pointer (per the
spec), it just hangs.

Dealing with dynamically sized arrays of classes appears to invoke the
same behavior, leading to seemingly random and implausible crashes &
bugs. But now that I am sticking with a single static instance of the
class, all the bugs are gone.

And yes, I thought I might be out of RAM, but if I stored the raw data
the class consumes in a gigantic array, then created static class
instances on demand whenever I wanted to access the data, no more bugs.
That takes way more memory (and CPU time) than the method I am using
now, so running out of memory was not the problem here.